2A202600032_DaoVanSon_Day16/requirements.txt
pydantic>=2.7
rich>=13.7
typer>=0.12
pandas>=2.2
python-dotenv>=1.0
google-generativeai>=0.5.2
openai>=1.0
